use nodes consensus with each other to pick in from waiting pool

+++ src/timeblock/time_block_manager.cc
@@ -115,7 +115,8 @@ bool TimeBlockManager::LeaderNewTimeBlockValid(uint64_t* new_time_block_tm) {
 
 bool TimeBlockManager::BackupheckNewTimeBlockValid(uint64_t new_time_block_tm) {
     auto now_tm = common::TimeUtils::TimestampSeconds();
-    if (abs((int64_t)new_time_block_tm - (int64_t)latest_time_block_tm_) < kTimeBlockTolerateSeconds) {
+    if (abs((int64_t)new_time_block_tm - (int64_t)latest_time_block_tm_) <
+            (int64_t)kTimeBlockTolerateSeconds) {
         return true;
     }
